refactor to use pure Page Object Model design
