--- src/test/java/org/openelisglobal/qaframework/automation/test/TestData.java
@@ -1,12 +1,0 @@
-package org.openelisglobal.qaframework.automation.test;
-
-public class TestData {
-
-	private static final String DEFAULT_PWD = "adminADMIN!";
-	public static class UserInfo {
-		public String username;
-		public String password = DEFAULT_PWD;
-		public String userId;
-	}
-
-}
